+++ Sources/SentryCrash/Recording/Monitors/SentryCrashMonitor_Signal.c
@@ -26,6 +26,7 @@
 //
 
 #include "SentryCrashMonitor_Signal.h"
+#include "SentryCrashCPU.h"
 #include "SentryCrashID.h"
 #include "SentryCrashMachineContext.h"
 #include "SentryCrashMonitorContext.h"
@@ -67,6 +68,29 @@ static char g_eventID[37];
 #    pragma mark - Callbacks -
 // ============================================================================
 
+/** Invoke previously installed signal handlers, adapted from Sentry Native:
+ * https://github.com/getsentry/sentry-native/blob/9895a5c3ffab4e59e0c8020484cdd5dbc648666d/src/backends/sentry_backend_inproc.c#L59-L76
+ */
+static void
+invokePreviousSignalHandlers(int sigNum, siginfo_t *signalInfo, void *userContext)
+{
+    const int *fatalSignals = sentrycrashsignal_fatalSignals();
+    int fatalSignalsCount = sentrycrashsignal_numFatalSignals();
+
+    for (int i = 0; i < fatalSignalsCount; ++i) {
+        if (fatalSignals[i] == sigNum) {
+            struct sigaction *handler = &g_previousSignalHandlers[i];
+            if (handler->sa_flags & SA_SIGINFO) {
+                handler->sa_sigaction(sigNum, signalInfo, userContext);
+            } else if (handler->sa_handler != SIG_DFL && handler->sa_handler != SIG_IGN) {
+                // This handler can only handle to signal number (ANSI C)
+                void (*func)(int) = handler->sa_handler;
+                func(sigNum);
+            }
+        }
+    }
+}
+
 /** Our custom signal handler.
  * Restore the default signal handlers, record the signal information, and
  * write a crash report.
@@ -83,6 +107,29 @@ static void
 handleSignal(int sigNum, siginfo_t *signalInfo, void *userContext)
 {
     SENTRY_ASYNC_SAFE_LOG_DEBUG("Trapped signal %d", sigNum);
+
+    if (sentrycrashcm_isManagedRuntime()) {
+        // Let managed Mono/CoreCLR runtime handle the signal first,
+        // as they may convert it into a managed exception.
+        SENTRY_ASYNC_SAFE_LOG_DEBUG(
+            "Detected managed Mono/CoreCLR runtime. Passing signal to previous handlers.");
+
+        uintptr_t sp = sentrycrashcpu_stackPointerFromUserContext(userContext);
+        uintptr_t ip = sentrycrashcpu_instructionAddressFromUserContext(userContext);
+
+        invokePreviousSignalHandlers(sigNum, signalInfo, userContext);
+
+        // If the stack or instruction pointer changed, the managed runtime
+        // converted the signal into a managed exception and changed the context.
+        // https://github.com/dotnet/runtime/blob/6d96e28597e7da0d790d495ba834cc4908e442cd/src/mono/mono/mini/exceptions-arm64.c#L538
+        if (sp != sentrycrashcpu_stackPointerFromUserContext(userContext)
+            || ip != sentrycrashcpu_instructionAddressFromUserContext(userContext)) {
+            SENTRY_ASYNC_SAFE_LOG_DEBUG(
+                "Signal converted to a managed exception. Aborting signal handling.");
+            return;
+        }
+    }
+
     if (g_isEnabled) {
         thread_act_array_t threads = NULL;
         mach_msg_type_number_t numThreads = 0;
@@ -110,9 +157,13 @@ handleSignal(int sigNum, siginfo_t *signalInfo, void *userContext)
         sentrycrashmc_resumeEnvironment(threads, numThreads);
     }
 
-    SENTRY_ASYNC_SAFE_LOG_DEBUG("Re-raising signal for regular handlers to catch.");
-    // This is technically not allowed, but it works in OSX and iOS.
-    raise(sigNum);
+    // Re-raise the signal to invoke the previous handlers unless already called
+    // above for managed runtimes.
+    if (!sentrycrashcm_isManagedRuntime()) {
+        SENTRY_ASYNC_SAFE_LOG_DEBUG("Re-raising signal for regular handlers to catch.");
+        // This is technically not allowed, but it works in OSX and iOS.
+        raise(sigNum);
+    }
 }
 
 // ============================================================================